requirements.txt
jieba==0.42.1
wordcloud==1.9.3
python-docx==1.0.1
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
matplotlib==3.8.3
openpyxl==3.1.2 